packages/pillbug-matrix/pyproject.toml
[project]
name = "pillbug-matrix"
version = "0.1.0"
description = "Matrix channel plugin for Pillbug"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillbug",
    "matrix-nio>=0.24",
]

[project.scripts]
pillbug-matrix-access-token = "pillbug_matrix.access_token:entrypoint"

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[tool.uv.sources]
pillbug = { workspace = true }

[tool.uv.build-backend]
module-name = "pillbug_matrix"
module-root = "src"
